refactor directory and code
